# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Valuz Agent API
  version: "0.3.0"
  description: Valuz Agent API contract — single source of truth
servers:
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/settings/preferences:
    get:
      summary: Get user personalization preferences
      operationId: getPreferences
      tags: [settings]
      responses:
        "200":
          description: Current personalization preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesResponse"
    patch:
      summary: Update user personalization preferences
      operationId: patchPreferences
      tags: [settings]
      description: |
        Only supplied fields are updated. Conversation citations default on;
        citation verification defaults off because it can add latency and model
        usage. Verification requires citations. Disabling citations also
        disables verification.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesPatch"
      responses:
        "200":
          description: Updated personalization preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesResponse"

  # ── Memory (inspect / prune / toggle the agent memory system) ────────
...
  /v1/memory/settings:
    patch:
      summary: Toggle the memory system on/off
      operationId: patchMemorySettings
      tags: [memory]
      description: |
        ``enabled`` is the master switch (injection + background extractor);
        ``auto_extract`` gates only the background extractor. Only sent keys
        are updated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySettingsPatch"
      responses:
        "200":
          description: Updated toggles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorySettings"
...
components:
...
  schemas:
...
    PreferencesResponse:
      type: object
      required:
        - default_timezone
        - default_locale
        - detected_timezone
        - theme
        - font_size
        - conversation_citations_enabled
        - conversation_verification_enabled
        - conversation_task_coverage_enabled
      properties:
        default_timezone: {type: string}
        default_locale: {type: string}
        detected_timezone: {type: string}
        theme: {type: string}
        font_size: {type: string}
        conversation_citations_enabled:
          type: boolean
          default: true
          description: Render canonical citation indices when source evidence is available.
        conversation_verification_enabled:
          type: boolean
          default: false
          description: Run claim and citation quality verification before publishing a reply.
        conversation_task_coverage_enabled:
          type: boolean
          default: true
          description: Check explicit task completeness and repair bounded omissions before publishing a reply.
    PreferencesPatch:
      type: object
      properties:
        default_timezone: {type: string, nullable: true}
        default_locale: {type: string, nullable: true}
        theme: {type: string, nullable: true}
        font_size: {type: string, nullable: true}
        conversation_citations_enabled: {type: boolean}
        conversation_verification_enabled: {type: boolean}
        conversation_task_coverage_enabled: {type: boolean}
...
    MemorySettings:
      type: object
      required: [enabled, auto_extract, custom_instructions]
      properties:
        enabled: {type: boolean}
        auto_extract: {type: boolean}
        custom_instructions: {type: string}
    MemorySettingsPatch:
      type: object
      properties:
        enabled: {type: boolean}
        auto_extract: {type: boolean}
        custom_instructions:
          type: string
          description: Global reviewer guidance; capped server-side (1500 chars).
...